Remove storage_proto dependency from storage_api
After having changed the storage_api to only expose Rust domain
objects, we no longer need the storage_proto dependency on
storage_api. Instead, storage_proto becomes an implementation detail
of storage_rocksdb.
