+++ src/common/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod errors;
 pub mod partitioner;
 pub mod retry_policy;
 pub mod traits;
