+++ src/storage_proto/Cargo.toml
@@ -9,7 +9,7 @@ publish = false
 
 [features]
 default = []
-conversion = ["dep:common", "dep:bytes", "dep:bytestring", "dep:opentelemetry_api", "dep:thiserror"]
+conversion = ["dep:common", "dep:bytes", "dep:bytestring", "dep:opentelemetry_api"]
 
 [dependencies]
 bytes = { workspace = true, optional = true }
@@ -18,7 +18,6 @@ common = { workspace = true, optional = true }
 opentelemetry_api = { workspace = true, optional = true }
 prost = { workspace = true }
 prost-types = { workspace = true }
-thiserror = { workspace = true, optional = true }
 
 [build-dependencies]
 prost-build = { workspace = true }
